requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.6.4
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.8
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
config==0.5.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
databricks-sdk==0.65.0
debugpy==1.8.16
decorator==5.2.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.2
fastmcp==2.12.3
Flask==3.1.2
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.14.0
mdurl==0.1.2
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
narwhals==2.5.0
nest-asyncio==1.6.0
numpy==2.3.3
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.4
pyperclip==1.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.38.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
wandb==0.22.0
wcwidth==0.2.13
Werkzeug==3.1.1
xgboost==3.0.5
zipp==3.23.0
